+++ lib/alpha_card/errors/invalid_attribute_type.rb
@@ -6,7 +6,7 @@ module AlphaCard
     # wrong attribute value type.
     #
     # @param value [Object] current attribute value
-    # @param values [Array] possible attribute types
+    # @param types [Array] possible attribute types
     def initialize(value, types)
       super("'#{value}' must be an instance of #{types.join(', ')}")
     end